refactor: move test files to src/__tests__
