+++ package.json
@@ -28,10 +28,10 @@
     "tslint:fix": "npm run tslint -- --fix",
     "pretest": "npm run clean && npm run build && npm run docker:start",
     "pretest:ci": "npm run build",
-    "test": "lb-mocha --allow-console-logs \"dist/test\"",
-    "test:ci": "lb-mocha --allow-console-logs \"dist/test\"",
+    "test": "lb-mocha --allow-console-logs \"dist/src/__tests__/**/*.js\"",
+    "test:ci": "lb-mocha --allow-console-logs \"dist/src/__tests__/**/*.js\"",
     "posttest": "npm run lint",
-    "test:dev": "lb-mocha --allow-console-logs dist/test/**/*.js && npm run posttest",
+    "test:dev": "lb-mocha --allow-console-logs dist/src/__tests__/**/*.js && npm run posttest",
     "prestart": "npm run build",
     "start": "concurrently --kill-others \"npm run start:app\" \"npm run start:recommender\"",
     "start:app": "node .",
