+++ src/__tests__/acceptance/ping.controller.acceptance.ts
