+++ tsconfig.json
@@ -3,9 +3,7 @@
   "extends": "./node_modules/@loopback/build/config/tsconfig.common.json",
   "include": [
     "src",
-    "test",
-    "recommender",
-    "index.ts"
+    "recommender"
   ],
   "exclude": [
     "node_modules/**",
